Molecules now being contained inside hollow sphere through collisions

+++ trunk/ideal-gas-v3/test-src/TestMoleculeHollowSphereCollisionII.java
@@ -84,7 +84,7 @@ public class TestMoleculeHollowSphereCollisionII extends PhetApplication {
                 public void actionPerformed( ActionEvent e ) {
                     m[0] = new HeavySpecies( new Point2D.Double( sphere1.getPosition().getX() + 37,
                                                                  sphere1.getPosition().getY() ),
-                                             new Vector2D.Double( 100, 130 ),
+                                             new Vector2D.Double( 100, 0 ),
                                              new Vector2D.Double(),
                                              5 );
 //                    m[0] = new HeavySpecies( new Point2D.Double( 383.3, 389.35 ),
